category

+++ src/main/java/com/hotmart/models/db/Category.java
@@ -22,11 +22,9 @@ public class Category implements Serializable{
 
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
 	private String name;
     
-    
-
 }
